+++ Source/BuildPackages/Templates/Console_contains.pas
@@ -1,0 +1,3 @@
+contains
+  NovusCommandLine in '..\..\core\Console\NovusCommandLine.pas',
+  NovusConsole in '..\..\core\Console\NovusConsole.pas';